remove html-asset-loader, preserve static/assets folder instead (#885)
This commit removes the html-asset-loader, which was used to automagically
perform webpack loads for static assets. As we have continued to add more complexity
to the theme, the scope of the html-asset-loader has grown to a difficult to maintain state.
The loader has to search through not only HTML, but also all javascript strings we provide.

Furthermore, the implementation of this loader is dependent on implementation specific
details of the html-loader.

Finally, removing this loader will make it much easier to migrate to ESBuild, or any other
bundler we choose.

I moved html-loader to the prod build only, and turned on all the default minification settings
instead of our custom ones. I also updated the version to the latest.
Also update browserlist (`npx browserslist@latest --update-db`) because it was asking to be updated.

J=SLAP-1419
TEST=manual

see that the test-site still displays static image assets correctly