use webpack asset modules instead of file-loader, fix preload
This commit replaces file-loader with webpack asset modules,
which is the webpack 5 way to do it. This is slightly more streamlined
than using loaders that are separate npm packages, which have been
deprecated by asset modules.

This commit also fixes font preloads no longer pointing correctly to font
assets. Originally, I tried to separate that out to another PR, however
it seems like html-loader@2 only works properly with asset modules and
not file-loader. As opposed to downgrading html-loader, then reupgrading
it in this PR, they're in the same PR.

I removed the minimize=true config for html-loader, because by default
html-loader will minimize when mode="production" and won't minimize when
mode="development"

J=SLAP-1084,SLAP-1419
TEST=manual

see that html is minified in prod mode, but not for dev mode

see that the font preloads get their URLs updated correctly,
for example pointing to source-sans.[hash].woff instead of
static/assets/fonts/source-sans.woff

see that bundle.css points to the correct URL, and that the
font shows up properly on the page