Add HitchhikerCTA-icon classes back (#899)
Add back the HitchhikerCTA divs and classes that were removed in #874

J=SLAP-1473
TEST=visual

Inspect the DOM and see the div and class present for the standard card. Inspect the updated Percy snapshots and see some of the icons change size slightly due to the CSS targeting the div. Also see the icon appear for the standard card which was previously missing.