add the custom-init-config partial
I considered adding this in HitchhikerJS, which
would give us better error messages if the a Hitchhiker messes
up JS syntax. On the other hand Rose asked for a partial,
and I'm a little worried about confusing Hitchhikers.
On the other hand probably only advanced Hitchhikers will use this.

In the future I was thinking jambo could pass a getter for
templates and partials into templatedatavalidator, or even
the whole HBS instance, so we can easily validate template/partials.

J=SLAP-1447
TEST=manual,auto

add a custom-init-config in the test site, with functions for onVerticalSearch
and onUniversalSearch