Use color-text-primary for card body text (#966)
Use the color-text-primary variable for the color of text on the body of cards

This change makes most text slightly lighter because the value of --yxt-color-text-primary is `#212121;`. Previously the cards were falling back to the default of the browser which is black.

The item mentioned using the text mixin to accomplish this, however I decided to use the variable directly because there is less chance of side effects because the text mixin affects things other than color.

J=SLAP-1617
TEST=manual

For testing purposes, set the text to red and view all of the cards of the test site. See that the color of the body of text is red as expected.