Fix double query on load from within iframe integrations (#986)
Fix double query on load from iframe integrations

When the full-page-map loads within an iframed experience, the iframe resizer resizes the page which leads to an additional query being fired on map load because it triggers the searchOnMapMove. Disabling `searchOnMapMove` would prevent the double query from firing, however it would cause the 'Search this area' button to appear which isn't supposed to happen until the map is moved by the user. 

By ignoring the first map change invocation from within iframe experiences, we can eliminate the issues described above.

J=SLAP-1633
TEST=manual

Load the vertical full page map from within an iframe and confirm from the network tab that only a single query is fired on page load. Test that when `disableSearchOnMapMove` is true, the 'Search This Area' button doesn't pop up until the user moves the map. Also test the pages with standard non-iframe integrations and check that the functionality is the same.