add vertical command: handle page and template namess with spaces in them
When calling spawn sync, in order for an argument with a space in it to be
recognized as a single argument, it must be wrapped in quotes. Otherwise, the
later parts of the argument will be considered unknown arguments.
For example, if I run

`npx jambo [COMMAND] --name Blog Article --template vertical-standard`

the "Article" part of Blog Article will be considered a separate argument from
--name=Blog. Instead, it needs to be

`npx jambo [COMMAND] --name 'Blog Article' --template vertical-standard`

I also added an error message if a user tries to use a verticalKey or page name that contains
a quote. This was not working previously, I wouldn't want it to be ambiguous as to whether this was
ever supported.

J=TECHOPS-2502
TEST=manual

test adding a vertical, with and without a space
verify that the arguments passed to the add vertical command in CBD through the SGS UI is
the same as the way I tested locally