Add null check for url formatter (#990)
Add a check for a null `baseUrl` passed to the `getUrlWithTextHighlight` formatter. Without this check, document search direct answer cards would display links that didn't exist.

J=SLAP-1667
TEST=auto, manual

Check that jest tests pass and that there is no 'read more' link shown at the bottom of document search cards if the baseUrl is null or undefined.