Update image formatter to use dynamic thumbnailer (#998)
Change the image formatter to use dynamic thumbnailer instead of expecting an array of thumbnails. A thumbnail array will no longer be provided after the streams migration. Dynamic thumbnailer uses the host `dynl` for images at least as large as the specified size (the smallest image with at least one dimension greater than or equal to the specified values). The host `dynm` gives the largest image that is smaller than the specified values in both dimensions.

Note: for `dynl`, desired size values of 1 or smaller are a special case which indicates that dimension should be ignored and the returned thumbnail should be larger than the other dimension.

J=SLAP-1670
TEST=auto, manual

Check that the current profile images still appear as before and those in the streams test account now appear as well.