Update feedback text (#1001)
Update the text of the feedback cards to "Thanks!"

Add translations as well. Note: the feedback text of the direct answer component stays the same. Since the "Thanks!" text is shorter, it fixes the long text shifting issue. If someone sets a custom text which is longer, then the text wrapping will appear, however that is the desired behavior according to product

J=SLAP-1698
TEST=manual

View the updated text in the theme test site in english and in Spanish