add missing prop comments when merging global configs on upgrade
Fixes an issue where new prop comments would not be added to a user's
global config on upgrade. This issue occurred due to how comment-json
stores comments in CommentToken arrays, under a symbol determined by the
closest json property, such as `Symbol.for('before:sessionTracking)`
When a comment-json assign is done on two CommentJSONValue-s that both
have comments under the same symbol, the json with higher priority will override
the other, as opposed to magically merging the arrays together with no duplicates.

This was fixed by re-adding in any missing comments.
No guarantees are made about the exact locations of re-added comments, to avoid
risk associated with a larger rework. I re-added them to the front for now, because
that works the best for the specific issue pointed out ("token" missing").

J=SLAP-1694
TEST=manual,auto

jambo upgrade, saw token added to my global config