add missing prop comments when merging global configs on upgrade (#1004)
Fixes an issue where new prop comments would not be added to a user's
global config on upgrade. This issue occurred due to how comment-json
stores comments in CommentToken arrays. Each CommentToken[] is associated
with a symbol determined by the closest json property, such as `Symbol.for('before:sessionTracking)`.
During a comment-json assign, CommentToken arrays associated with the same symbol
are not magically merged with no duplicates. Instead, the higher priority json will override the other.

This was fixed by re-adding in any missing comments.
No guarantees are made about the exact locations of re-added comments, to avoid
risk associated with a larger rework. I re-added them to the front for now, because
that works the best for the specific issue pointed out ("token" missing").

J=SLAP-1694
TEST=manual,auto

jambo upgrade, saw token added to my global config