Card layout updates (#1005)
Update the layout of the card feedback buttons so that they are located in their own row underneath the CTAs and the card content

This change was in response to feedback from Product/UX. This PR moves the feedback buttons out from the `HitchhikerCard-actions` div into the card body. We also remove the `HichhikerCard-actions` div since that was only there so that the buttons could be next to the CTAs. This allows the card content to be full width when CTAs are not present. It adds some extra height to most cards since there's an entire row dedicated to the feedback buttons, however UX was happy with this decision. As part of this change, I removed some styling which is no longer relevant.

This PR also ensures that the feedback buttons are in the bottom right on grid layout. (Previously the buttons could appear just below the card content in grid layout just below the CTAs)

Also set a fixed feedbackContent height so that the cards don't shift when the buttons are pressed.

J=SLAP-1702
TEST=manual

Visual inspect all cards with feedback buttons enabled and also check mobile layouts. Test the product cards on the grid layout and confirm the feedback buttons are in the bottom right. Test clicking the feedback buttons and observing that the cards do not shift