Fix distance alignment (#1006)
Fix distance alignment for financial-professional-location and professional-location cards so that distance is on the far right of the top row of the card instead of right next to the title. Also, fix distance alignment with the close card button on vertical full page map for both cards and location-standard cards so that distance is in-line to the left of the button.

J=SLAP-1718
TEST=manual

Check that both professional location cards (and multilang) have correct distance alignment, matching the location-standard card. Check that mobile view of vertical full page map has correct alignment of distance with close card button for location-standard card