Remove scrollbars from filters (#1015)
Change styling of `yxt-FilterOptions-options` to remove scrollbars that were appearing on filters even when all options were displayed for some cases with custom font sizes. Add a `--yxt-filters-and-sorts-line-height` variable so that customers can change the line-height of the `optionLabel` easily if the scrollbar appears with even larger custom font sizes.

J=SLAP-1727
TEST=visual

See that filter scrollbar was removed on a site where it was previously appearing.