Fix handling of `universalLimit` in `verticalsToConfig`. (#1021)
Previously, if a `universalLimit` were specified for a vertical, we would perform client-side filtering. Instead, we should be creating a `search.universalLimit` object and passing that to `ANSWERS.init`. Not only does this remove responsibility from the front-end, but it allows the HH to display more than 10 results per vertical on the Universal page.

This PR adds a `getDefaultUniversalLimit` helper that iterates through the vertical page configs and uses the `verticalsToConfig.universalLimits` to build a `search.universalLimit`. 

J=SLAP-1749
TEST=manual

Tested the following scenarios and verified correct behavior in each:

- No `universalLimit` values are set and there is no limit specified in the `global_config` or `pageSettings`.
- Some `universalLimit` values are set, but not others. There is no universal limit in the `global_config` or `pageSettings`.
- Some `universalLimit` values are set, but a universal limit is also present in the `global_config`.