Integrate with Yext Sites (#1035)
Integrate with Yext Sites
 
Add config to support yext sites. The main "gotcha" is that Sites renames the answers-hitchhiker-theme to 'repo', so I added an extra jambo config which we can use with Yext Sites which refers the the theme as 'repo'.

Once we merge this into develop and master, we should be able to see those sites hosted at:
- theme.slapshot.pagescdn.com
- https://develop-theme-slapshot-pagescdn-com.preview.pagescdn.com/

The live preview integration is a little out of scope, but I think it would be useful if we add it in the future. I ran into an issue with it which may also have to do with repo renaming. It would allow us to open up the test site for any PR that we push to Github which would be useful for testing

J=SLAP-1679
TEST=manual

Connect this branch to Yext Sites and see the site hosted at https://devserveonyextsites-theme-slapshot-pagescdn-com.preview.pagescdn.com/