Add multilang doc search DA card (#1041)
Add a multi-lang document search DA card. Update index.es.json and index.ar.json to use this card for featured snippets in the test-site. Fix a bug in the allfields DA card template that affected the styling for the title when there is an icon.

J=SLAP-1830
TEST=manual

Check that the doc search card has translated feedback text and that the allfields DA card title is styled correctly with an icon.