Change ID used for map pins (#1049)
Use a result's internal `uid` instead of the external entity `id` for identifying map pins. The `id` can be any string specified by the user. This causes an error when clicking on a map pin on mobile if the `id` has spaces, apostrophes, etc. The `uid` is a unique, automatically-generated string that only consists of numbers, so it would not have this problem.

J=SLAP-1962
TEST=manual

Check that map pins on mobile can now be clicked regardless of external entity ID because the `uid` is used instead.