fix Map errors when the map container does not have offsetWidth/Height
If the full-page-map's container has an offsetWidth or offsetHeight of 0,
which essentially means that the map is not visible, the page will throw an error
and not render the map. This PR updates the map to gracefully handle these cases.

T=4215
TEST=manual

saw that the user's experience on the aforementioned techops ticket
no longer logs `'NaN' must be convertible to 'Number'`, and that
changing the page's styling so that the map becomes visible renders the map