Fix CollapsibleFilters tab order (#1057)
In the vertical page templates, move the View Results button from the results header to after the filters so that for collapsible filters the tab order goes through the filters first before going to the button at the bottom of the page. Update the page patch files for the test-site to reflect the changed page templates.

I also looked into doing this using the `tabindex` attribute, as specifying values greater than 0 sets the tab ordering of a page. However, this is not recommended for accessibility reasons because screen reader announcements don't follow `tabindex` order. Also, it would be cumbersome to add and maintain because each element on the page would need to be assigned a `tabindex` to maintain a natural order.

J=SLAP-829
TEST=manual

Check various vertical pages in the test-site (including people, healthcare_professionals, locations, and locations_full_page_map_with_filters) and see that the tab ordering is now correct.