add truncatedDetails to showMoreDetails to support rtf truncation
This commit adds a new truncatedDetails property to the showMoreDetails
portion of dataForRender's return value. This property is intended to be used
for truncating rich text card details, though it is not necessarily limited to that.

Updated all cards that use ANSWERS.formatRichText by default to have truncatedDetails
instead of showMoreLimit in their showMoreDetails. We are not removing showMoreLimit.

J=SLAP-2039
TEST=manual,auto

updated the products test-site page to have rtf truncation