Fix error resolving dependencies (#1072)
Bump the `grunt-webpack` version in `static/package.json` to fix the error about resolving dependencies. v4.0.3 supports v5 of `webpack` as a peer dependency, which is what we use in the package.json. It seems like the inconsistency may have been introduced when upgrading to `lockfileVersion` 2.

![image](https://user-images.githubusercontent.com/88398086/170762266-5a8f331d-4166-47fc-b3f1-6acf4507a899.png)

J=none
TEST=manual

Set up the test-site and smoke tested. Saw that the error no longer occurred when running setup-test-site.