Add English translations (#1077)
Add an English PO file so that English is treated like all other languages. This ensures that translations are still made correctly when there is a fallback locale specified for 'en'. Previously, when this was the case, the fallback locale's translations would be used because no English PO file was found by Jambo. By adding this translation file, even if a user adds custom text in a non-English language, the fallback and default translations will now be handled consistently and correctly.

J=SLAP-2127
TEST=manual

In the test-site `locale_config`, specified a fallback locale of 'es' for 'en' and saw that, before the changes, text was always translated into Spanish. After adding the PO file, English translations were used unless a match wasn't found, in which case the Spanish translation was correctly used as the fallback.