Fix full page map keyboard on android, and set current center correctly (#1101)
This PR fixes a bug where, on Android, it was not possible to run searches on the full page map, because
clicking into the keyboard would trigger a "searchOnMapMove", which would cause new search results to come in
and cause the keyboard to collapse itself.

The first issue is that, the android keyboard shifts all content on the page up, unlike the iOS keyboard, which does
not affect page content. This would cause the center of the map to move, which would cause searchOnMapMove
to be triggered. This was dealt with by fixing the height of the mapWrapper on page load for android devices.

The second issue was that the VerticalFullPageMapOrchestrator would ALWAYS have the incorrect map center.
This is because it would grab the center of the map before the map had actually loaded, resulting in it thinking the
center of the map was always the `this.defaultCenter` value, when it is virtually never that. This would make
it so that even with the first issue being patched, the keyboard would still always collapse itself the first time
you clicked into it, since the VerticalFullPageMapOrchestrator would not have the correct center until after that.
To patch this I added a new loadHandler to the map, and update the center in that handler.

J=TECHOPS-6738
TEST=manual

test on browserstack android chrome that I can click into the searchbar on the full page map, for both google and mapbox, and no search will be triggered and the keyboard will not hide itself and refuse to be typed into

test that locally on desktop the page looks fine on both google and mapbox