Remove Browserstack acceptance testing on IE11 (#1112)
Now that the Google Maps JS API has dropped support for IE11, Browserstack acceptance tests on IE11 fail. This PR updates the Browserstack acceptance workflow to only run the jobs on Safari and Firefox, and not on IE11 anymore. The `IE11NoCacheHook` is also removed as it's no longer needed.

J=SLAP-2502
TEST=auto

See that there are only Browserstack checks for Safari and Firefox on the `Test` commit in this PR and neither has Google Maps API errors.