dateRange formatter: Use dateFormatOptions for same-day end times
If an end date is on the same day, the dateRange formatter will
not show the month, day, and year of the end date, only the time.

However, it was passing in the default format options with
month, day, and year removed instead of using the format
options passed in through the dateFormatOptions parameter.

This update makes it so that dateFormatOptions are respected
for end times with only the options for month, day, and year
removed.

J=PC-201542
TEST=manual