Upgrade Simple Git to latest major and minor. (#1131)
This PR upgrades the version of `simple-git` used by the Theme. This will address a number of Snyk issues. In the latest version of `simple-git`, you import `simple-git` directly, not `simple-git/promise`. 

TEST=manual

Ensured that the logic for determining if the Theme is a submodule still worked as expected. This is the only usage of `simple-git` in the Theme.