Support EU Cloud Region
Add a cloud region config option to the global_config to allow
either 'us' (default), or 'eu' to be set as the cloud region.

The cloud region determines where the search assets and ytag
are sourced from.

J=BACK-2277
TEST=manual

Using the test site, set the cloudRegion to eu and confirmed
the assets are loaded from the eu cdn. Currently there is a
cors error, but I will retest them once Aether fixes it.